CircuitBreaker-grpc/fastapi/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.86.0
grpcio==1.50.0
grpcio-tools==1.50.0
h11==0.14.0
idna==3.4
protobuf==4.21.9
pydantic==1.10.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.19.0